Fixing example for web
Changing dart:io to universal_io

dart:io is not supported in web applications. universal_io is a cross-platform drop-in replacement for dart:io.

Changed two instances of the usage of dart:io in fullstack_example, and added universal_io to the pubspec of fullstack_example.

The example app now works correctly on the web.